pyproject.toml
[project]
name = "rag-master-book"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb==0.5.23",
    "faiss-cpu>=1.13.0",
    "flashrank>=0.2.10",
    "kiwipiepy>=0.23.1",
    "langchain>=0.3.0,<0.4.0",
    "langchain-chroma",
    "langchain-community>=0.3.0,<0.4.0",
    "langchain-core>=0.3.84",
    "langchain-openai>=0.3.0,<0.4.0",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters>=0.3.0,<0.4.0",
    "langgraph>=1.0.1",
    "pypdf>=6.10.0",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.4.0",
    "tavily-python>=0.7.23",
    "tiktoken>=0.12.0",
]
